+++ CHANGELOG.md
@@ -16,3 +16,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix erring ember try in ember-canary, properly remove component element wrapper.
 - Cleaned up code syntax to match airbnb's javascript style guide and eslint.
+- When clearing filter of relationships, dot notation was not maintained.